Floats and class vars to constants
I decided that it was better programming practice (looking at the very
clean go.rb and others) to have the keywords, pseudokeywords, etc. as
constants rather than as class variables under Nim.  (It also makes the
documentation for Nim.rb much cleaner).

Handling floats:
In Pygments, we can push multiple states on the stack (and then yield a
token) by creating a rule with something like:

<state name> (if needed):
  r'<regex>', <token name>, ['state1','state2']

But in Rouge, you can only have a rule with one of:
  rule /regex/,token
  rule /regex/,token,state
  rule /regex/, { |match| #block }

so to push both and then yield Num::Float, I created a block that pushed
both (and then yielded Num::Float with token).  Will do testing to see
if this works.