Fix for number suffixes
I edited the regexes for the integers and floats to better support
suffixes.  In the case of floats, I had an erroneous "\'" (escaped
single quote) in the regex for the 'Numbers' block because of how the
Pygments regex was setup.

For unsigned integers, I added an extra rule in the ':intsuffix' state
to check for 'u' or 'U'. This allows the annotation for unsigned
integers as specified in the manual.