Optional apostrophe w/ numeric types
This fixes the issue described here:
https://github.com/Araq/Nimrod/wiki/Pygments-issues in the section on
'Incorrect type suffix highlighting':

var x = 0f64
var y = 0'f64

should have the same highlighting with/without apostrophe.  The only
case in which the apostrophe is *not* optional is when specifying type
precision with hexadecimal digits - that *can* cause ambiguity, so the
apostrophe is required. Numeric regexes updated to accomodate this.