Examples for Nim
Added examples in the lib/rouge/demos/ and spec/visual/samples/
directories for the Nim programming language.  The former is a (somewhat
contrived) example of fixed width output and mathematical functions; the
other is taken from http://nimrod-by-example.github.io/, which is
available under public domain per terms of its license as of time of
writing.